+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-## v0.3.1 - To be released
+##  To be released
 
+* Added a new shape: Rectangle
 * Upgraded to Rust 2018 edition
 
 ## v0.3.0 - 2018-11-04